create `ArrayAccessTrait`
