+++ src/Trait/ArrayIteratorTrait.php
@@ -11,7 +11,7 @@ use IteratorAggregate;
  * Trait to implement {@see IteratorAggregate} interface for ActiveRecord.
  *
  * @method array getAttributes(array|null $names = null, array $except = [])
- * @see ActiveRecordInterface::getAttributes() for more info.
+ * @see ActiveRecordInterface::getAttributes()
  */
 trait ArrayIteratorTrait
 {
